finish apply page
