+++ src/commonStyles/variables.scss
@@ -10,8 +10,7 @@ $button-transparent-shadow-color: #14565A;
 $button-disabled: #C2D2D2;
 
 $input-border: $button-disabled;
+$input-hover: $button-hover-color;
 $dropdown-select-hover: #F6F6F0;
-
-$inputs-hover: $button-hover-color;
 
 $scrollbar: $button-disabled;
